Add the slug-generator to the Location model
Add a slug generator to the Location model also added to the save method of the model.
When a Location instance is created, a slug is generated, based on the designation field.
If the slug is already taken, one or two digits are appended to the end of the slug until the slug becomes unique.