Add missing fields and datacheck for reservations
Add time, price and status fields to Reservation model.
Add migration files ready to migrate.
Add Representation.save() which calculate the number of available seats
of a representation based on the total number of seats and the already
booked seats of all its reservations flagged as "Completed".
Add Reservation.save() which calculate the total price of the
reservation based on the unit price of the show and the amount of seats
booked. The method also updates the Reprensentation.